Refactor: refactored the code for improved readability and performance

+++ src/main/java/br/com/ourplaces/our_places_api/repository/CoupleRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface CoupleRepository extends JpaRepository<Couple, Integer> {
+public interface CoupleRepository extends JpaRepository<Couple, Long> {
     Optional<Couple> findByInviteCode(String inviteCode);
 
     Optional<Couple> findByUser1IdOrUser2Id(Long user1Id, Long user2Id);
